% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{accuracy}
\alias{accuracy}
\title{Accuracy measures for a regression models}
\usage{
accuracy(actual, pred)
}
\arguments{
\item{actual}{Actual data as numeric vector}

\item{pred}{Predicted values as numeric vector}
}
\description{
Accuracy measures for a regression models
}
\examples{
\dontrun{

accuracy(acc, pred)}


}
\author{
Resul Akay
}
