\name{Blanciforti86}
\docType{data}
\alias{Blanciforti86}

\title{Artificial Prices and Quantities}

\description{
   The \code{Blanciforti86} data frame contains data on food demand in
   the U.S.A. from 1947 to 1978. These data include total food expenditure,
   expenditure shares of 4 food commodity groups and price indices for
   these 4 food commodity groups. The 4 groups are 1 = meats, 2 = fruits
   and vegetables, 3 = cereal and bakery products, and 4 = miscellaneous foods.
}

\usage{data(Blanciforti86)}

\format{
   This data frame contains the following columns:
   \describe{
      \item{pFood1}{Price index of food commodity group 1.}
      \item{pFood2}{Price index of food commodity group 2.}
      \item{pFood3}{Price index of food commodity group 3.}
      \item{pFood4}{Price index of food commodity group 4.}
      \item{wFood1}{Expenditure share of food commodity group 1 in
         total food expenditure.}
      \item{wFood2}{Expenditure share of food commodity group 2 in
         total food expenditure.}
      \item{wFood3}{Expenditure share of food commodity group 3 in
         total food expenditure.}
      \item{wFood4}{Expenditure share of food commodity group 4 in
         total food expenditure.}
      \item{xFood}{Expenditure for Food.}
   }
}

\source{
   Blanciforti, Laura. A., Richard D. Green and Gordon A. King (1986)
   U.S. Consumer Behavior Over the Postwar Period:
   An Almost Ideal Demand System Analysis.
   Giannini Foundation Monograph Number 40, August 1986.
}

\keyword{datasets}
