\name{Bleymueller251}
\docType{data}
\alias{Bleymueller251}

\title{Artificial Prices and Quantities}

\description{
   The \code{Bleymueller251} data frame contains prices and quantities of
   4 products for the years 1970, 1974 and 1978.
   This data are part of Exercise 25.1 of Bleymueller, Gehler und Guetlicher (1979).
}

\usage{data(Bleymueller251)}

\format{
   This data frame contains the following columns:
   \describe{
      \item{p.A}{Price of good A.}
      \item{q.A}{Quantity of good A.}
      \item{p.B}{Price of good B.}
      \item{q.B}{Quantity of good B.}
      \item{p.C}{Price of good C.}
      \item{q.C}{Quantity of good C.}
      \item{p.D}{Price of good D.}
      \item{q.D}{Quantity of good D.}
   }
}

\source{
   Bleymueller, J; G. Gehlert and H. Guelicher (1979)
   Statistik für Wirtschaftswissenschaftler. Verlag Vahlen, Muenchen.
}

\keyword{datasets}
