% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equivalent_variation.R
\name{equivalent_variation}
\alias{equivalent_variation}
\title{Equivalent variation}
\usage{
equivalent_variation(he_m, he_p, alpha, Y_0)
}
\arguments{
\item{he_m}{Market rent}

\item{he_p}{Program rent}

\item{alpha}{Rent/income}

\item{Y_0}{Income}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}

Equivalent variation (EV) is a measure of economic welfare changes
associated with changes in prices. John Hicks (1939) is attributed
with introducing the concept of compensating and equivalent variation.
The equivalent variation is the change in wealth, at current prices,
that would have the same effect on consumer welfare as would the change in
prices, with income unchanged. It is a useful tool when the present prices
are the best place to make a comparison.
}
\examples{
\dontrun{
equivalent_variation(he_m, he_p, alpha, Y_0)
}

}
\author{
Resul Akay
}
