% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get_var_imp}
\alias{get_var_imp}
\title{Catboost variable importance}
\usage{
get_var_imp(object)
}
\arguments{
\item{object}{A catboost model}
}
\value{
A data frame class of \code{varimp, data.frame)}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}
}
\examples{

get_var_imp(fit)

}
\author{
Resul Akay
}
