\name{micEcon-internal}
\alias{checkNames}
\alias{micEconIndex}
\alias{print.snqProfitImposeConvexity}
\alias{snqProfitCoef}
\alias{snqProfitCoefNames}
\alias{snqProfitElaJacobian}
\alias{snqProfitRestrict}
\alias{snqProfitSystem}
\alias{snqProfitTestCoef}
\alias{translogMonoRestr}
% Document the following:

%%%%


\title{ Internal micEcon Functions }

\description{
   Internal micEcon Functions
}

\details{
   These are either various methods, or functions, not to be called
   directly by the user (or in some cases are just
   waiting for proper documentation to be written :).
}

\keyword{ internal }
