\name{Missong03E7.7}
\docType{data}
\alias{Missong03E7.7}

\title{Meat Prices and Quantities in Germany}

\description{
   The \code{Missong03E7.7} data frame contains meat prices and demanded quantities
   of a representative (West-)German household for the years 1986 to 1989.
   This data are part of Exercise 7.7 of Missong (2003).
}

\usage{data(Missong03E7.7)}

\format{
   This data frame contains the following columns:
   \describe{
      \item{p.beef}{Average price of beef (DM/kg).}
      \item{q.beef}{Demanded Quantity of beef (kg).}
      \item{p.veal}{Average price of veal (DM/kg).}
      \item{q.veal}{Demanded Quantity of veal (kg).}
      \item{p.pork}{Average price of pork (DM/kg).}
      \item{q.pork}{Demanded Quantity of pork (kg).}
   }
}

\source{
   Missong, M. (2003)
   Aufgabensammlung zur deskriptiven Statistik, Oldenbourg, Muenchen.

   Statistisches Bundesamt (1989)
   Fachserie 15, Reihe 1, p. 76f.
}

\keyword{datasets}
