% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{plot.varimp}
\alias{plot.varimp}
\title{Catboost variable importance plot}
\usage{
\method{plot}{varimp}(varimp)
}
\arguments{
\item{varimp}{A dataframe class of varimp, e.g. from get_var_imp}
}
\value{
variable importance plot
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}
}
\examples{

varimp <- get_var_imp(fit)

plot.varimp(varimp)

}
\author{
Resul Akay
}
