% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.auto_catboost.R
\name{predict.auto_catboost}
\alias{predict.auto_catboost}
\title{Predict an auto_catboost object}
\usage{
\method{predict}{auto_catboost}(object, newdata, ...)
}
\arguments{
\item{object}{An auto_catboost object.}

\item{newdata}{The newdata}

\item{\dots}{Ignored for now}
}
\description{
Predict an auto_catboost object
}
\author{
Resul Akay
}
