\name{residuals.translogEst}
\alias{residuals.translogEst}

\title{Residuals of a Translog function}

\description{
   Extract the residuals from the estimation of a Translog function.
}

\usage{
\method{residuals}{translogEst}( object, \dots )
}

\arguments{
   \item{object}{an object of class \code{translogEst}.}
   \item{\dots}{currently not used.}
}

\value{
   \code{residuals.translogEst} returns a vector
   containing the residuals of an estimated translog function.
}

\author{Arne Henningsen \email{ahenningsen@agric-econ.uni-kiel.de}}


\seealso{
   \code{\link{translogEst}} and \code{\link{residuals}}
}

\keyword{methods}
