% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{split_data}
\alias{split_data}
\title{Split data randomly in to train and test.}
\usage{
split_data(data, p = 0.75, set_seed = 24)
}
\arguments{
\item{data}{A data.frame.}

\item{p}{percentage of train data.}

\item{set_seed}{Random seed}
}
\description{
Split data randomly in to train and test.
}
\examples{
\dontrun{
dta <- split_data(data)
}

}
