\name{testConsist}
\alias{testConsist}
\title{Testing Theoretical Consistency}
\description{
   Test theoretical consistency of microeconomic models.
}

\usage{
testConsist( object, ... )
}

\arguments{
  \item{object}{a microeconomic model}
  \item{\dots}{further arguments for methods}
}

\details{
  This is a generic function.
}

\author{Arne Henningsen \email{ahenningsen@agric-econ.uni-kiel.de}}

\seealso{\code{\link{testConsist.aidsEst}}}

\keyword{methods}
