\name{micEcon-internal}
\alias{aidsCoef}
\alias{aidsElasJacobian}
\alias{aidsJacobian}
\alias{aidsRestr}
\alias{aidsSystem}
\alias{checkNames}
\alias{coef.summary.probit}
\alias{micEconIndex}
\alias{model.frame.probit}
\alias{model.matrix.probit}
\alias{print.aidsTestConsist}
\alias{print.snqProfitImposeConvexity}
\alias{print.summary.maximisation}
\alias{print.summary.maxLik}
\alias{print.summary.probit}
\alias{returnCode.maximisation}
\alias{returnCode.maxLik}
\alias{snqProfitCoef}
\alias{snqProfitCoefNames}
\alias{snqProfitElaJacobian}
\alias{snqProfitRestrict}
\alias{snqProfitSystem}
\alias{snqProfitTestCoef}
\alias{summary.probit}
\alias{translogMonoRestr}
% Document the following:

%%%%


\title{ Internal micEcon Functions }

\description{
   Internal micEcon Functions
}

\details{
   These are either various methods, or functions, not to be called
   directly by the user (or in some cases are just
   waiting for proper documentation to be written :).
}

\keyword{ internal }
