\name{micEcon-internal}
\alias{aidsCoef}
\alias{aidsJacobian}
\alias{aidsRestr}
\alias{aidsSystem}
\alias{coef.summary.probit}
\alias{micEconIndex}
\alias{model.frame.probit}
\alias{model.matrix.probit}
\alias{print.aidsEst}
\alias{print.aidsTestConsist}
\alias{print.heckit}
\alias{print.snqProfitImposeConvexity}
\alias{print.summary.maximisation}
\alias{print.summary.maxLik}
\alias{print.summary.probit}
\alias{print.summary.tobit2}
\alias{checkNames}
\alias{snqProfitCoef}
\alias{snqProfitCoefNames}
\alias{snqProfitRestrict}
\alias{snqProfitSystem}
\alias{snqProfitTestCoef}
\alias{summary.probit}
\alias{summary.tobit2}
\alias{translogMonoRestr}
\alias{vcov.maxLik}

\title{ Internal micEcon Functions }

\description{
   Internal micEcon Functions
}

\details{
   These are either various methods, or functions, not to be called
   directly by the user (or in some cases are just
   waiting for proper documentation to be written :).
}

\keyword{ internal }
